u32 to f32 for const

+++ src/infos.rs
@@ -1,11 +1,11 @@
 
-pub const FIELD_WIDTH: u32 = 182;
-pub const FIELD_DEPTH: u32 = 243;
-pub const SPACE_BEFORE_LINE_SIDE: u32 = 12;
-pub const GOAL_WIDTH: u32 = 60;
-pub const DISTANCE_YGOAL_FROM_CENTER: u32 = 115;
-pub const ROBOT_RADIUS: u32 = 9;
-pub const BALL_RADIUS: u32 = 2;
-pub const ENBUT_DEPTH: u32 = 30;
-pub const ENBUT_WIDTH: u32 = 75;
-pub const ENBUT_RADIUS: u32 = 10;
+pub const FIELD_WIDTH: f32 = 182.0;
+pub const FIELD_DEPTH: f32 = 243.0;
+pub const SPACE_BEFORE_LINE_SIDE: f32 = 12.0;
+pub const GOAL_WIDTH: f32 = 60.0;
+pub const DISTANCE_YGOAL_FROM_CENTER: f32 = 115.0;
+pub const ROBOT_RADIUS: f32 = 9.0;
+pub const BALL_RADIUS: f32 = 2.0;
+pub const ENBUT_DEPTH: f32 = 30.0;
+pub const ENBUT_WIDTH: f32 = 75.0;
+pub const ENBUT_RADIUS: f32 = 10.0;
